Add files via upload
Change ToDouble

+++ Program.cs
@@ -12,9 +12,9 @@
             Console.Write("Your full name is "+fname+" "+lname);
 
             Console.Write("Enter a number: ");
-            int num1 = Convert.ToInt32(Console.ReadLine());
+            double num1 = Convert.ToDouble(Console.ReadLine());
             Console.Write("Enter another number: ");
-            int num2 = Convert.ToInt32(Console.ReadLine());
+            double num2 = Convert.ToDouble(Console.ReadLine());
 
             Console.WriteLine(num1 + num2);
